# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Documentation",
        "description": "L5 Swagger",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0."
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the user by the provided ID with associated posts.",
                "operationId": "9d01b92f8b824e6ac3f636baedf023c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to retrieve",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "account_id": {
                                                    "type": "integer",
                                                    "example": "4"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Bùi Thúy Ngọc"
                                                },
                                                "sex": {
                                                    "type": "string",
                                                    "example": "female"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Hà Đông, Hà Nội, Việt Nam"
                                                },
                                                "birthday": {
                                                    "type": "string",
                                                    "example": "2005-12-03"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-20T09:43:36.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-20T09:43:36.000000Z"
                                                },
                                                "posts": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "7"
                                                            },
                                                            "content": {
                                                                "type": "string",
                                                                "example": "post mới nè"
                                                            },
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": "2"
                                                            },
                                                            "latitude": {
                                                                "type": "number",
                                                                "example": null
                                                            },
                                                            "longitude": {
                                                                "type": "number",
                                                                "example": null
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-03-30T17:07:19.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-03-30T17:07:19.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get user successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer )",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}